// nestery-backend/src/features/loyalty/enums/loyalty-transaction-type.enum.ts
export enum LoyaltyTransactionTypeEnum {
  BOOKING_COMMISSION_EARN = 'BOOKING_COMMISSION_EARN',
  REFERRAL_BONUS_EARN = 'REFERRAL_BONUS_EARN',
  REVIEW_AWARD_EARN = 'REVIEW_AWARD_EARN',
  DAILY_CHECKIN_EARN = 'DAILY_CHECKIN_EARN',
  PROFILE_COMPLETION_EARN = 'PROFILE_COMPLETION_EARN',
  PREMIUM_SUBSCRIPTION_BONUS_EARN = 'PREMIUM_SUBSCRIPTION_BONUS_EARN',
  PARTNER_OFFER_EARN = 'PARTNER_OFFER_EARN',
  PREMIUM_DISCOUNT_REDEEM = 'PREMIUM_DISCOUNT_REDEEM',
  TEMP_FEATURE_ACCESS_REDEEM = 'TEMP_FEATURE_ACCESS_REDEEM',
  PROFILE_CUSTOMIZATION_REDEEM = 'PROFILE_CUSTOMIZATION_REDEEM',
  PRIZE_DRAW_ENTRY_REDEEM = 'PRIZE_DRAW_ENTRY_REDEEM',
  PARTNER_SERVICE_DISCOUNT_REDEEM = 'PARTNER_SERVICE_DISCOUNT_REDEEM',
  ADJUSTMENT_ADD = 'ADJUSTMENT_ADD',
  ADJUSTMENT_SUBTRACT = 'ADJUSTMENT_SUBTRACT',
}